add codecs, frame rate, mime type info from parsed hls info

+++ lib/src/asms/better_player_asms_track.dart
@@ -13,7 +13,7 @@ class BetterPlayerAsmsTrack {
   final int? bitrate;
 
   ///Frame rate of the track
-  final int? frameRate;
+  final double? frameRate;
 
   ///Codecs of the track
   final String? codecs;
